requirements.txt
flask 
Flask-SQLAlchemy
flask-restful
newspaper3k